third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/form-submission-0/form-data-set-empty-file.window-expected.txt

This is a testharness.js-based test.
FAIL Empty <input type=file> is still serialized assert_equals: type expected "application/octet-stream" but got ""
Harness: the test ran to completion.
